pyproject.toml
[tool.poetry]
name = "travel-agent"
version = "0.1.0"
description = ""
authors = ["BenderScript <6779302+BenderScript@users.noreply.github.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.3"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.4"
amadeus = "^10.0.0"
langchainhub = "0.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
